refactor: rename component file
